requirements.txt
# Chinese → English Document Translator Dependencies
numpy==1.26.4
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
paddleocr==2.7.0.3
paddlepaddle==2.6.2
pdf2image==1.17.0
streamlit==1.32.0
pillow==10.2.0

# FIXED Gemini deps
google-generativeai==0.5.4
google-api-core==2.17.0
grpcio==1.60.0
